data-service/requirements.txt
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
APScheduler>=3.10.0
requests>=2.31.0
google-genai>=0.1.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
yfinance>=0.2.36
